pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "git-substitute"
version = "0.1.0"
description = ""
authors = ["Robert Roos <robert.soor@gmail.com>"]
readme = "README.md"
packages = [
    { include = "git_substitute", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.43"

[tool.poetry.scripts]
git_substitute = "git_substitute.__main__:main"
